This checkin in not quite complete in that it looks as if it has a bad interaction with sstools, but I want to upload it and then hunt down that effect. There is now the basic infrastructure to allow almost all CSL vectors to be of essentially arbitrary size, rather than being limite to fitting within 4Mbyes which was the previous situation. I need to work all through the code both to check it and to rationalize code that previously used lower levels of abstraction. The consequence of the change that is most immediate is that I have been able to simplify the object-vector (ie package system) hash tables in a way that allows them to cope with very very many symbols in a way better than was the situation up until rather recently. The new policy is that the C++ inline function "elt" accesses memberfs of a general vector, which is either a single block of memory or a 2-level structure with the top level marked as TYPE_INDEXVEC. The function "basic_elt" copes with just the simple linear block. I have provided inline functions to do a range of tests on vectors allowing for th epossibility of index blocks - I now need to migrate the code so as to use them everywhere!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4146 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/sysfwin.cpp
@@ -873,7 +873,7 @@ const char *CSLtmpdir()
 #endif
 }
 
-const char *CSLtmpnam(const char *suffix, int32_t suffixlen)
+const char *CSLtmpnam(const char *suffix, size_t suffixlen)
 {   time_t t0 = time(NULL);
     clock_t c0 = clock();
     unsigned long taskid;
